Separate loggers on their own module
- Also, add some of @dbkeitel 's suggestions:
   - `get_version_string` rather than importing from version.
   - Do not populate directory if user doesn't ask for it.
