+++ examples/run_all_examples.py
@@ -2,7 +2,8 @@ import os
 import shutil
 from glob import glob
 
-from pyfstat.helper_functions import run_commandline, set_up_logger
+from pyfstat import set_up_logger
+from pyfstat.helper_functions import run_commandline
 
 exit_on_first_failure = False
