+++ pyfstat/__init__.py
@@ -16,11 +16,11 @@ from .grid_based_searches import (
     TransientGridSearch,
 )
 from .gridcorner import gridcorner
-from .helper_functions import set_up_logger
 from .injection_parameters import (
     AllSkyInjectionParametersGenerator,
     InjectionParametersGenerator,
 )
+from .loggers import set_up_logger
 from .make_sfts import (
     BinaryModulatedWriter,
     FrequencyAmplitudeModulatedArtifactWriter,
